+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## [0.0.1] - Initial release
 
 Added:
-- Everything
+- Initial implementation with theme, controller and of course, the splash screen widget.